avr-evtd: Cast number of commands to an int.
We need to find a better way to map between strings and indices and that is
not costly in terms of execution.

+++ avr-evtd.cpp
@@ -879,7 +879,7 @@ static void parse_config(char *content)
 			}
 
 			/* Locate our expected commands */
-			for (cmd = 0; cmd < NCOMMANDS; cmd++)
+			for (cmd = 0; cmd < (int) NCOMMANDS; cmd++)
 				if (strcasecmp(pos, command[cmd]) == 0)
 					break;
